Created the initial UIElementsUtility class.
Created UIElementsUtility.cs
It holds utility functions specific to Sprite Editor work flows. Common UI Utility functions will be in the SF UI Elements package.

## Added

### Vector2 GetTextureDisplaySize() function
This function and it's overloads help get the display size of textures of the pass in Visual Element and either a passed in Texture2d or Sprite.

### float PixelsPerUnitScaleForElement()
Gets the pixels per unit of the sprite used in a Visual Element based on the referenced visual element panels scaled pixels per point.