Created the SF Sprite Editor Settings Provider.
The SF Sprite Editor Settings Provider is used to create a set of project settings in Unity settings menu.

## Added

### SFSPriteEditorSettingsProvider
The SFSpriteEditorSettingsProvider class creates the UI for the project settings menu and retrieves the values from the SFSpriteEditorSettings scriptable object to allow users to edit the values. It also adds the ability to use the search bar to find certain key words in it.

### SFSpriteEditorSettings
This is the scriptable object that will hold the data that is read and set in the project settings provider class.