Created the Sprite Data classes.
## Added

### Sprite Data
Sprite Data inherits from Sprite Rect and is used to keep track of the data textures being spliced into a sprite sheet. It was created to add tile and animation customization potential to the normal Sprite Rect.

###  Sprite Data Cache
This is a scriptable object used to keep track of changes done in the Sprite Editors to Sprite Rects and Sprites before applying the changes to the actual Texture2Dfile that the sprites are sliced from.

It has helper functions for bulk updating sprites data like names, pivot points, and so forth.