Created the a default SpriteData Provider for SpriteNameFileIDPairs.
## Added

### DefaultSpriteNameFileIDProvider
We needed a DefaultSpriteNameFileIDProvider with it's implementation to be used when any null value comes back from trying to get a ISpriteEditorDataProvider. This way if we are trying to get a data provider from a sprite rect that hasn't had one set yet we can add a new one and use it.

This method will be used for TileDataProviders as well when the TileDataProvider class is needed.