Created the start of the ViewRectZoomManipulator for UI Elements.
Added the ViewRectZoomManipulator that allows for Zooming in on a UI Element.

TODO: More tweaks are needed for helping multiple View Rect Matrix values. Such as changes to a View Rect Matrix using TRS.

TODO: Needing to make sure this works for IMGUI Containers events as well. This could work by having a C# Action that is invoked when zooming in and calls a function that updates IMGUI Containers.